Version 1.2.1. Replaced incorrect "greedy" match_wxhlx() that returned the longest helix in the interval [minlen,maxlen] with the correct algorithm that returns a vector of ALL helices in the interval.  This also required enclosing the inner search of find_wchlx, find_pknot3 and and find_4plex in a loop that ranges over all of these helices.
Modified the input language to accept single quoted strings 'xxx'.  These
strings uninterpreted save for the C escapes.  Double quoted are now always
interpreted as sequences, with upper case forced to lower, u->t and the iupac
codes replaced by the appropriate RE.

Powered up =~ and !~ operators in score.

Restricted grammar to use the in operator only in a site:

	de():de...de() in { "a:u", ..., "a:a" }

as opposed to expr in expr.  This operator is not yet functional as it
core dumps somewhere.

Finally the new match_wchlx() finds more helices in certain cases and so
copied the test file over the corresponding chk files.