Version 1.4.0:  Includes. Done via #include stmts in the source.  Can be nested.  Implemented via an initial pass over the source writing the included source into a temp file which is then close and reopened and passed to lex/yacc.  Line/file info is saved in this temp file as # line N F directives.
A fair amount of detail work was required to update error reporting as the
error(s) can be in separate files.  This went double for runtime (ie score)
error(s) as a simple stack is insufficient to do this.

Finally, by using a true preprocess include phase, I was able to reimplement
the cmd line parm defs so that they are done immediately before processing
the res. word `descr'

CVs: ----------------------------------------------------------------------