Fixed mrnamotif to be able to correctly handle the -v, -s switches: Give the info and quit without searching.  Also fixed logic so that -v, -s, -c switches do not require data.
Added 2 new switches to mrnamotif: -pre cmd, -post cmd.  If present, these
are inserted before/after the worker node rnamotif cmd:

	pre-cmd -fmt X data-file | rnamotif ... | post-cmd

Note that the pre-cmd gets both the fmt option and the data file.  If the
pre cmd is present, rnamotif reads from stdin.

The idea was to allow users to do some worker node post processing to
cut down on the output from prelim. searches.  Suggested by Lesley Collins

Also moved all prototypes of all visible functions in the rnamotif source
to rnamot.h (or dbutil.h).

This required making the yacc generated parser 64 bit clean, which it wasn't
because the default parse stack type was int.  Now a union of int/NODE_T *,
so no longer requires that int/pointer be the same.  Required a careful
rereading of the yacc manual to get this right!